backend/geo-agent/requirements.txt
langchain
langchain-google-genai
pymongo
requests
python-dotenv
schedule
